Test adjustments

+++ app/controllers/api/v1/cart_items_controller.rb
@@ -2,6 +2,8 @@ class Api::V1::CartItemsController < ApplicationController
   before_action :authenticate_user!
   before_action :set_cart
 
+  rescue_from ActiveRecord::RecordNotFound, with: :record_not_found
+
   def create
     product = Product.find(params[:product_id])
     item = @cart.cart_items.find_or_initialize_by(product: product)
@@ -59,32 +61,41 @@ class Api::V1::CartItemsController < ApplicationController
 
   def destroy
     item = @cart.cart_items.find(params[:id])
-    item.destroy
 
-    total_price = @cart.cart_items.includes(:product).sum { |cart_item| cart_item.product.price * cart_item.quantity }
-    @cart.update!(total_price: total_price)
+    if item
+      item.destroy
+      total_price = @cart.cart_items.includes(:product).sum { |cart_item| cart_item.product.price * cart_item.quantity }
+      @cart.update!(total_price: total_price)
 
-    render json: {
-      id: @cart.id,
-      total_price: @cart.total_price,
-      cart_items: @cart.cart_items.map do |cart_item|
-        {
-          id: cart_item.id,
-          quantity: cart_item.quantity,
-          product: {
-            id: cart_item.product.id,
-            name: cart_item.product.name,
-            price: cart_item.product.price,
-            product_picture_url: cart_item.product.product_picture.attached? ? Rails.application.routes.url_helpers.rails_blob_url(cart_item.product.product_picture, only_path: true) : nil
-          }
-        }
-      end
-    }
+      render json: { id: @cart.id, total_price: @cart.total_price, cart_items: @cart.cart_items }, status: :ok
+      # render json: {
+      #   id: @cart.id,
+      #   total_price: @cart.total_price,
+      #   cart_items: @cart.cart_items.map do |cart_item|
+      #     {
+      #       id: cart_item.id,
+      #       quantity: cart_item.quantity,
+      #       product: {
+      #         id: cart_item.product.id,
+      #         name: cart_item.product.name,
+      #         price: cart_item.product.price,
+      #         product_picture_url: cart_item.product.product_picture.attached? ? Rails.application.routes.url_helpers.rails_blob_url(cart_item.product.product_picture, only_path: true) : nil
+      #       }
+      #     }
+      #   end
+      # }
+    else
+      render json: { error: "Item não encontrado" }, status: :not_found
+    end
   end
 
   private
 
   def set_cart
     @cart = current_user.cart || current_user.create_cart
+  end
+
+  def record_not_found
+    render json: { error: "Recurso não encontrado" }, status: :not_found
   end
 end
